# docs/openapi.yaml
openapi: 3.0.3
info:
  title: El Greco API
  version: 1.0.0
  description: |-
    REST API for the El Greco web application.

    • Authentication: cookie-based (HttpOnly `access_token`) with optional Bearer token support.
    • CSRF: double-submit cookie. For POST/PUT/PATCH/DELETE include header `X-CSRF-Token` that matches `csrf_token`.
    • Auth state: `GET /auth/state` returns `{ authenticated, user? }` with HTTP 200.
    • Protected resources: under `/api/*` and require authentication (cookie or bearer).

    Uploads use `multipart/form-data`. Listing endpoints support pagination and basic text search via `q`.
servers:
  - url: http://localhost:4000
    description: Local (auth endpoints)
  - url: http://localhost:4000/api
    description: Local (API endpoints)
  - url: https://your-backend-domain.com
    description: Production (auth endpoints)
  - url: https://your-backend-domain.com/api
    description: Production (API endpoints)
...
paths:
...
  /paintings/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Paintings]
      summary: Get a painting by id
      responses:
        '200':
          description: A painting
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Painting'
        '404':
          $ref: '#/components/responses/NotFoundError'
    patch:
      tags: [Paintings]
      summary: Update a painting (optionally replace image)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                year:
                  type: string
                description:
                  type: string
                categoryId:
                  $ref: '#/components/schemas/ObjectId'
                subcategoryId:
                  $ref: '#/components/schemas/ObjectId'
                image:
                  type: string
                  format: binary
            encoding:
              image:
                contentType: image/jpeg, image/png, image/webp
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Painting'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      tags: [Paintings]
      summary: Delete a painting
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /links:
    get:
      tags: [Links]
      summary: List links (external resources)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/QParam'
        - in: query
          name: paintingId
          schema:
            $ref: '#/components/schemas/ObjectId'
          description: Filter by related painting id
      responses:
        '200':
          description: A paginated list of links
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedLinkList'
    post:
      tags: [Links]
      summary: Create a link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkCreateInput'
      responses:
        '201':
          description: Created link
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Link'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /links/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Links]
      summary: Get a link by id
      responses:
        '200':
          description: A link
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Link'
        '404':
          $ref: '#/components/responses/NotFoundError'
    patch:
      tags: [Links]
      summary: Update a link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkUpdateInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Link'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      tags: [Links]
      summary: Delete a link
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        $ref: '#/components/schemas/ObjectId'
      description: Resource identifier
    PageParam:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number (1-based)
    LimitParam:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Page size
    SortParam:
      in: query
      name: sort
      schema:
        type: string
      description: Sort expression, e.g. `title` or `-createdAt`
    QParam:
      in: query
      name: q
      schema:
        type: string
      description: Simple text search on name/title/description fields
  responses:
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            default:
              value:
                error: unauthorized
                message: Missing or invalid token
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            default:
              value:
                error: not_found
                message: Resource not found
    BadRequestError:
      description: Bad request / validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          examples:
            default:
              value:
                error: validation_error
                message: Invalid input
                details:
                  - path: title
                    message: title is required
  schemas:
    ObjectId:
      type: string
      pattern: ^[a-fA-F0-9]{24}$
      example: 60f718d9c2b9a60012abc123
...
    Painting:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/ObjectId'
        title:
          type: string
        year:
          type: string
        description:
          type: string
        categoryId:
          $ref: '#/components/schemas/ObjectId'
        subcategoryId:
          $ref: '#/components/schemas/ObjectId'
        imageUrl:
          type: string
          description: Public URL to the stored image (if using object storage) or API route to fetch image
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Link:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/ObjectId'
        paintingId:
          $ref: '#/components/schemas/ObjectId'
        title:
          type: string
        url:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LinkCreateInput:
      type: object
      required: [title, url]
      properties:
        paintingId:
          $ref: '#/components/schemas/ObjectId'
        title:
          type: string
        url:
          type: string
          format: uri
    LinkUpdateInput:
      type: object
      properties:
        paintingId:
          $ref: '#/components/schemas/ObjectId'
        title:
          type: string
        url:
          type: string
          format: uri
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        totalDocs:
          type: integer
          example: 42
        totalPages:
          type: integer
          example: 5
        hasNextPage:
          type: boolean
        hasPrevPage:
          type: boolean
...
    PaginatedLinkList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Link'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            details:
              type: array
              items:
                type: object
                properties:
                  path:
                    type: string
                  message:
                    type: string

# Global default security (all paths require auth unless they override with security: [])
security:
  - cookieAuth: []
    bearerAuth: []